testing: update pipeline build tests; move assertion helper fn to shared tests utils file
